Initial commit of MCMC analysis directory
This directory contains the core trace data used as input into the analysis and the analysis R script itself.

The analysis uses a a local cloned directory for file input/output, so the current working directory in the R script may need to be changed depending on who's running it.

I'm working on updating the script that generates the figures from the analysis output. I'll add that ASAP

Co-Authored-By: MPatterson5 <34280680+MPatterson5@users.noreply.github.com>